+++ packages/designer/src/components/widgets/user-avatar/index.vue
@@ -10,9 +10,18 @@
   import { ElAvatar } from 'element-plus';
   import { UserFilled } from '@vtj/icons';
   import { useEngine } from '../../../framework';
+  import { useCheckVersion } from '../../hooks';
+  import { notify } from '../../../utils';
 
   const engine = useEngine();
   const { access } = engine.adapter || {};
+
+  useCheckVersion((version) => {
+    notify(
+      `VTJ 发布了版本 ${version}，您可以通过重新安装依赖更新到最新版本`,
+      '版本更新'
+    );
+  });
 
   const avatarSrc = computed(() => {
     const avatar = access?.getData()?.avatar || '';