+++ platforms/h5/src/index.ts
@@ -1,0 +1,3 @@
+export * from './modules';
+export * from './renderer';
+export * from './utils';