+++ platforms/pro/ide/views/page.vue
@@ -1,23 +1,25 @@
 <template>
-  <XMask>
-    <component v-if="renderer" :is="renderer"></component>
-  </XMask>
+  <component v-if="renderer" :is="renderer"></component>
 </template>
 <script lang="ts" setup>
   import { ref, getCurrentInstance } from 'vue';
   import { useRoute } from 'vue-router';
-  import { XMask } from '@vtj/ui';
+  // import { XMask } from '@vtj/ui';
   import { ACCESS_STORAGE_KEY } from '../contants';
   import {
     createProvider,
     LocalService,
     ContextMode,
     Extension,
-    Access
+    Access,
+    createAdapter,
+    createServiceRequest
   } from '../../src';
   import { IconsPlugin } from '@vtj/icons';
   import { ElMessageBox } from 'element-plus';
-  const service = new LocalService();
+  import { notify, loading } from '../utils';
+  const adapter = createAdapter({ loading, notify });
+  const service = new LocalService(createServiceRequest(notify));
   const config = await service.getExtension().catch(() => null);
   const { options, adapters } = config
     ? await new Extension(config).load()
@@ -36,7 +38,7 @@
     mode: ContextMode.Runtime,
     service,
     materialPath: __BASE_PATH__,
-    adapter: { access, remote },
+    adapter: Object.assign(adapter, { access, remote }, options?.adapter || {}),
     ...(options || {}),
     dependencies: {
       Vue: () => import('vue'),