+++ platforms/pro/CHANGELOG.md
@@ -3,6 +3,171 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.26](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.25...@vtj/pro@0.9.26) (2025-02-11)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.25](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.24...@vtj/pro@0.9.25) (2025-01-23)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.24](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.23...@vtj/pro@0.9.24) (2025-01-22)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.23](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.22...@vtj/pro@0.9.23) (2025-01-22)
+
+
+### Bug Fixes
+
+* 🐛 被引用的区块渲染时总是刷新的问题 ([3b7a4e0](https://gitee.com/newgateway/vtj/commits/3b7a4e06e94085d3b9613c30cfe1b75bbeaf1a84))
+
+
+
+
+
+## [0.9.22](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.21...@vtj/pro@0.9.22) (2025-01-16)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.21](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.20...@vtj/pro@0.9.21) (2025-01-14)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.20](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.19...@vtj/pro@0.9.20) (2025-01-14)
+
+
+### Bug Fixes
+
+* 🐛 engine pageBasePath ([c516c01](https://gitee.com/newgateway/vtj/commits/c516c01dd9bc97681250db80c80694baad837e9e))
+
+
+
+
+
+## [0.9.19](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.18...@vtj/pro@0.9.19) (2025-01-14)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.18](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.17...@vtj/pro@0.9.18) (2025-01-14)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.17](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.16...@vtj/pro@0.9.17) (2025-01-14)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.16](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.15...@vtj/pro@0.9.16) (2025-01-13)
+
+
+### Features
+
+* ✨ h5平台 ([0497ccb](https://gitee.com/newgateway/vtj/commits/0497ccbb53f01a537176cae36f69b5d3019fc68e))
+* ✨ H5平台协议 ([51484a8](https://gitee.com/newgateway/vtj/commits/51484a8d0723897e462e1e2c37e16dbdf2546bea))
+
+
+
+
+
+## [0.9.15](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.14...@vtj/pro@0.9.15) (2025-01-13)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.14](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.13...@vtj/pro@0.9.14) (2025-01-10)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.13](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.12...@vtj/pro@0.9.13) (2025-01-10)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.12](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.11...@vtj/pro@0.9.12) (2025-01-10)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.11](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.10...@vtj/pro@0.9.11) (2025-01-09)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.10](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.9...@vtj/pro@0.9.10) (2025-01-09)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.9](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.8...@vtj/pro@0.9.9) (2025-01-08)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.9.8](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.7...@vtj/pro@0.9.8) (2025-01-06)
+
+
+### Features
+
+* ✨ @vtj/renderer 与 element-plus 解耦 ([6e63fe8](https://gitee.com/newgateway/vtj/commits/6e63fe81bf5e7047f1c621277b9104d2b44b02bd))
+
+
+
+
+
 ## [0.9.7](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.6...@vtj/pro@0.9.7) (2025-01-04)
 
 **Note:** Version bump only for package @vtj/pro